[AMA-022] - Organise utils folder module structure (#24)
* fix: create services folder for utils

* update: utils services imports

* feat: add cors to API

* fest: create utils decorators folder

* update: utils decorators imports

* feat: create utils filters folder

* update: utils filters imports

* feat: create utils interfaces folder

* update: utils interfaces imports

* feat: create utils common folder

* update: utils common imports

* update: rename utils types file
