+++ src/country/country.controller.ts
@@ -3,10 +3,10 @@ import { ApiTags } from '@nestjs/swagger';
 import { CountryService } from './country.service';
 import { CreateCountryDto, UpdateCountryDto } from './country.dto';
 import { CountryDocument } from './country.entity';
-import { PaginationOptions, PaginationResult } from '../utils/pagination.util';
+import { PaginationOptions, PaginationResult } from '../utils/common/pagination.util';
 import { UserRole } from '../auth/auth.types';
 import { RolesGuard } from '../auth/auth.roles.guard';
-import { Roles } from '../utils/roles.decorator.util';
+import { Roles } from '../utils/decorators/roles.decorator.util';
 
 @ApiTags('Countries')
 @UseGuards(RolesGuard)
