+++ src/main.ts
@@ -4,13 +4,14 @@ import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { INestApplication, Logger } from '@nestjs/common';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import { ExceptionsFilter } from './utils/exception.filter.util';
-import { ApiResponseUtil } from './utils/api.response.util';
-import { Environment } from './utils/types.util';
+import { ExceptionsFilter } from './utils/filters/exception.filter.util';
+import { ApiResponseUtil } from './utils/common/api.response.util';
+import { Environment } from './utils/utils.types';
 import { rateLimitMiddleware } from './utils/middlewares/rate-limit.middleware';
 import { RedisService } from './redis/redis.service';
 import { RedisCacheInterceptor } from './redis/redis.interceptor';
 import { ClusterService } from './cluster/cluster.service';
+import { IpWhitelistService } from './utils/services/ip.whitelist.service.util';
 
 function useSwagger(app: INestApplication) {
   const options = new DocumentBuilder()
@@ -24,14 +25,20 @@ function useSwagger(app: INestApplication) {
 async function bootstrap(): Promise<void> {
   const app = await NestFactory.create(AppModule);
   const env = process.env.ENVIRONMENT;
+  const ipWhitelistService = new IpWhitelistService;
 
   useSwagger(app);
+
   app.use(rateLimitMiddleware);
   app.useGlobalFilters(new ExceptionsFilter());
   app.useGlobalInterceptors(
     new RedisCacheInterceptor(new RedisService),
     new ApiResponseUtil()
   );
+  app.enableCors({
+    origin: ipWhitelistService.getAllowedIps(),
+    credentials: true
+  });
 
   if (env === Environment.PRODUCTION) {
     const httpsOptions = {
