+++ src/state/state.controller.ts
@@ -14,10 +14,10 @@ import { StateService } from './state.service';
 import { CreateStateDto, UpdateStateDto } from './state.dto';
 import { StateDocument } from './state.entity';
 import { ApiTags } from '@nestjs/swagger';
-import { PaginationOptions, PaginationResult } from '../utils/pagination.util';
+import { PaginationOptions, PaginationResult } from '../utils/common/pagination.util';
 import { UserRole } from '../auth/auth.types';
 import { RolesGuard } from '../auth/auth.roles.guard';
-import { Roles } from '../utils/roles.decorator.util';
+import { Roles } from '../utils/decorators/roles.decorator.util';
 
 @ApiTags('States')
 @UseGuards(RolesGuard)
