+++ src/utils/common/pagination.util.ts
@@ -1,6 +1,6 @@
 import { Model, Document, Query, FilterQuery } from 'mongoose';
 import { NotFoundException } from '@nestjs/common';
-import { FilterDto } from './util.dto';
+import { FilterDto } from '../util.dto';
 
 export type SortOrder = 'asc' | 'desc';
