+++ src/utils/decorators/roles.decorator.util.ts
@@ -1,4 +1,4 @@
 import { SetMetadata } from '@nestjs/common';
-import { UserRole } from '../auth/auth.types';
+import { UserRole } from '../../auth/auth.types';
 
 export const Roles = (...roles: UserRole[]) => SetMetadata('roles', roles);
