+++ src/utils/utils.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { ApiResponseUtil } from './api.response.util';
-import { ExceptionsFilter } from './exception.filter.util';
-import { ApiKeyService } from './api.key.service.util';
-import { IpWhitelistService } from './ip.whitelist.service.util';
+import { ApiResponseUtil } from './common/api.response.util';
+import { ExceptionsFilter } from './filters/exception.filter.util';
+import { ApiKeyService } from './services/api.key.service.util';
+import { IpWhitelistService } from './services/ip.whitelist.service.util';
 
 @Module({
   providers: [
